requirements.txt
absl-py==2.1.0
addict==2.4.0
albumentations==1.2.1
aliyun-python-sdk-core==2.15.1
aliyun-python-sdk-kms==2.16.3
asttokens==2.4.1
attrs==24.3.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.2.0
Brotli @ file:///home/conda/feedstock_root/build_artifacts/brotli-split_1695989787169/work
causal_conv1d @ file:///home/zhaoming/VMamba/causal-conv1d-main
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1720457958366/work/certifi
cffi @ file:///home/conda/feedstock_root/build_artifacts/cffi_1696001684923/work
chardet==5.2.0
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1698833585322/work
click==8.1.7
cloudpickle==3.0.0
cmake==3.30.0
colorama==0.4.6
ConfigArgParse==1.7
contourpy==1.2.1
crcmod==1.7
cryptography==42.0.8
cycler==0.12.1
Cython==3.0.10
decorator==5.1.1
defusedxml==0.7.1
docker-pycreds==0.4.0
docopt==0.6.2
einops==0.8.0
exceptiongroup==1.2.1
executing==2.0.1
fastjsonschema==2.21.1
filelock==3.14.0
fonttools==4.53.1
fsspec==2024.6.1
ftfy==6.2.0
fvcore==0.1.5.post20221221
gitdb==4.0.11
GitPython==3.1.43
gmpy2 @ file:///home/conda/feedstock_root/build_artifacts/gmpy2_1715527283764/work
grpcio==1.66.1
h2 @ file:///home/conda/feedstock_root/build_artifacts/h2_1634280454336/work
hpack==4.0.0
huggingface-hub==0.23.4
hyperframe @ file:///home/conda/feedstock_root/build_artifacts/hyperframe_1619110129307/work
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1713279365350/work
imageio==2.34.2
imageio-ffmpeg==0.5.1
importlib_metadata==8.0.0
iniconfig==2.0.0
iopath==0.1.10
ipdb==0.13.11
ipython==8.12.3
jedi==0.19.1
Jinja2 @ file:///home/conda/feedstock_root/build_artifacts/jinja2_1715127149914/work
jmespath==0.10.0
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
kiwisolver==1.4.5
lazy_loader==0.4
lit==18.1.8
mamba_ssm @ file:///home/zhaoming/VMamba/mamba
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe @ file:///home/conda/feedstock_root/build_artifacts/markupsafe_1706899921127/work
mat4py==0.6.0
matplotlib==3.8.4
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.0
mmcv==2.1.0
mmdet==3.3.0
mmengine==0.10.1
mmpretrain==1.2.0
mmsegmentation==1.2.2
model-index==0.1.11
modelindex==0.0.2
mpmath @ file:///home/conda/feedstock_root/build_artifacts/mpmath_1678228039184/work
nbclient==0.10.2
nbconvert==7.16.4
nbformat==5.10.4
networkx @ file:///home/conda/feedstock_root/build_artifacts/networkx_1712540363324/work
ninja==1.11.1.1
numpy==1.22.4
numpy-hilbert-curve==1.0.1
opencv-python==4.6.0.66
opencv-python-headless==4.10.0.84
opendatalab==0.0.10
openmim==0.3.9
openxlab==0.1.1
ordered-set==4.1.0
oss2==2.17.0
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1718189413536/work
pandas==1.4.3
pandocfilters==1.5.1
parso==0.8.4
pathlib==1.0.1
pathtools==0.1.2
pexpect==4.9.0
pickleshare==0.7.5
Pillow==9.1.1
pipreqs==0.5.0
platformdirs==4.2.2
pluggy==1.5.0
portalocker==2.10.0
prettytable==3.10.2
promise==2.3
prompt_toolkit==3.0.47
protobuf==3.20.3
psutil==6.0.0
ptyprocess==0.7.0
pure-eval==0.2.2
pycocotools==2.0.8
pycparser @ file:///home/conda/feedstock_root/build_artifacts/pycparser_1711811537435/work
pycryptodome==3.20.0
Pygments==2.18.0
pyparsing==3.1.2
PySocks @ file:///home/conda/feedstock_root/build_artifacts/pysocks_1661604839144/work
pytest==8.2.2
python-dateutil==2.9.0.post0
pytz==2023.4
PyYAML==6.0.1
pyzmq==26.2.0
pyzorder==0.0.2
qudida==0.0.4
referencing==0.35.1
regex==2024.5.15
requests==2.28.2
rich==13.4.2
rpds-py==0.22.3
safetensors==0.4.3
scikit-image==0.22.0
scikit-learn==1.5.1
scipy==1.13.1
seaborn==0.13.2
selective_scan @ file:///home/zhaoming/VMamba/VMamba-main/kernels/selective_scan
sentry-sdk==2.10.0
setproctitle==1.3.3
shapely==2.0.5
shortuuid==1.0.13
six==1.16.0
smmap==5.0.1
some-package==0.1
soupsieve==2.6
stack-data==0.6.3
submitit==1.5.1
sympy @ file:///home/conda/feedstock_root/build_artifacts/sympy_1718625539893/work
tabulate==0.9.0
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
termcolor==2.4.0
terminaltables==3.1.10
threadpoolctl==3.5.0
tifffile==2024.7.2
timm==0.4.12
tinycss2==1.4.0
tokenizers==0.19.1
tomli==2.0.1
torch==2.0.1
torchaudio==2.0.2
torchinfo==1.8.0
torchmetrics==0.9.3
torchvision==0.15.2
tornado==6.4.2
tqdm==4.64.0
traitlets==5.14.3
transformers==4.42.3
triton==2.1.0
ttach==0.0.3
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1717802530399/work
tzdata==2024.1
urllib3==1.26.19
wandb==0.12.21
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.0.4
yacs==0.1.8
yapf==0.40.2
yarg==0.1.9
zipp==3.19.2
zstandard==0.22.0
